Fixed linking errors

+++ SFMLBoids/src/classes/UIManager.cpp
@@ -14,7 +14,7 @@ void UIManager::pollEvent() {
 	while (window.pollEvent(event)) {
 		if (event.type == sf::Event::Resized)
 		{
-			sf::FloatRect visibleArea(0, 0, event.size.width, event.size.height);
+			sf::FloatRect visibleArea(0, 0, (signed int)event.size.width, (signed int)event.size.height);
 			window.setView(sf::View(visibleArea));
 		}
 		if (event.type == sf::Event::Closed) {
